+++ Backend/Authentication/AuthenticationResponse.cs
@@ -5,7 +5,7 @@ namespace AspTwitter.Authentication
 {
     public class AuthenticationResponse
     {
-        public long Id { get; set; }
+        public uint Id { get; set; }
         public string Name { get; set; }
         public string Username { get; set; }
         public string About { get; set; }
