+++ Backend/Requests/EntryRequest.cs
@@ -6,7 +6,7 @@ namespace AspTwitter.Requests
 {
     public class EntryRequest
     {
-        public long AuthorId { get; set; }
+        public uint AuthorId { get; set; }
         public string Text { get; set; }
     }
 }
